Fix all animations (#20)
* Refactor to all 'contour' langauge to 'path' and subpath language

* Get PoC kind of working for morphing between paths with different number of verts

* Get child parent relationship kind of working with animations

* Fix deleting animation segments on the timeline

* Fix deleting anim objects

* Get child parent relationships working

* Try to reduce nanovg draw calls (it looks like it didn't work :/)

* Add debug info to render passes

* Add serialization to some editor settings

* Start breaking text animations with new model

* Add typing to animId and animObjId and remove all pointer references to try to reduce errors when removing anim objects

* Fix nasty little bug in serialization (TODO add automatic serialization registration type thingy)

* Get text animations working pretty good

* Fixed one bug with replacement letters not matching source letters

* Add scene management (#19)

* Add basic scene manager serialization

* Get basic scene management working with scary crash

* Fix bug where we incorrectly tried to write to a buffer that was too small

* Make scene manager prettier

* Get persistent changes to individual letters

* Change anim obj ids to uint64_t

* Fix svg interpolation holes

* Fix one bug with interpolation

* Add rough draft of SVG cache

* Add a couple minor optimizations

* Break a lot of stuff and optimize nothing

* Kind of get LRU cache working

* Get it kind of working...

* Switch from snapshots to delta based change

* Fix some infinite loop bugs

* Get much faster outline drawing working

* Refactor 3d path rendering to 2d

* Make my life better

* Get the bevels working

* Switch other apis to use the new path api

* Get rid of snapshots and kind of fix gizmos

* Get changes propagating again

* Get create animation working a-OK again

* Get basic svg file objects working

* Add generic svg file support

* Get changes showing up immediately and clean up some memory leaks

* Add UI error messages when importing an SVG fails

* Fix bug from uninitialized global transforms

* Fix parser bug: off by one

* Final tries to fix this stuff for the night

* Remove stroke width and other now unnecessary code

* Remove special cases for svg groups while rendering

* Get svg groups layout correct and add some metrics

* Fix buggy offset with svg texture coordinates

* Fix svg holes breaking when reloading svg

* REMOVE NANOVG (FINALLY) And switch to PlutoVG which is sooo much better

* Remove nanovg submodule

* Update build scripts

* Fix some more layouting issues

* Trying to fix some stuff

* Fix layouting for more complex svg groups

* Add references to animations in anim objects

* Add animation references

* Add checkbox option to apply animations to children

* Stub out style parsing in svg paths

* Add basic css parser

* Add support for inline styles and stuff

* Make svg parsing more robust by fixing edge cases

* Add one more edge case

* Make the cameras regular anim objects

* Add fixed array size enforcement helper

* Fix move to animations

* Fix gizmo position bug

* Make windows moveable from titlebar only

* Add errors for unhandled switch cases

* Fix unhandled switch case statements

* Add beginning of circumscribe animation

* Add multicolored path support

* Add render outline method

* Fix some bugs and support path colors for outline rendering

* Add component-wise multiplication for vectors

* Begin fixing warnings

* Get the rest of the circumscribe animations working

* Add circle shape to the circumscribe animation

+++ Animations/src/audio/WavLoader.cpp
@@ -66,7 +66,7 @@ namespace MathAnim
 			res.sampleRate = header.dataHeader.sampleRate;
 
 			uint8* audioData = (uint8*)g_memory_allocate(sizeof(uint8) * header.dataHeader.dataSize);
-			uint8 amtRead = fread(audioData, sizeof(uint8) * header.dataHeader.dataSize, 1, fp);
+			size_t amtRead = fread(audioData, sizeof(uint8) * header.dataHeader.dataSize, 1, fp);
 			g_logger_assert(amtRead == 1, "Failed to read all the audio data.");
 			fclose(fp);
 			res.audioData = audioData;
